+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 - support for MFA to cognito logins
+- `authService.setupMfa()`
+- `authService.disableMfa()`
 
 ## [3.2.0] - 2023-05-08